Fix an unnervingly annoying problem
By specifying the main entry and the electron context directory in the package.json , electron got a little bit confused the paths specified were relative to the context directory which made it super annoying to call upon the html. Took in total around 8 hours to find the source of this unexpected behavior.